pkg/mount: move mountinfo filters to separate file
This is just cosmetic change.
